reorganize project structure
No reason to make rails deal with all those files in lib when
they're generated from src anyways.

Now we build to build/ and produce only the result in lib/ for rails
to find.

This should lead to less unknown in the future as to how rails will build
as well, since there's no node-to-browser magic going on inside the final
gem.